+++ rust/src/registers_methods.rs
@@ -33,6 +33,8 @@ extern "C" {
         [[*const core::ffi::c_char; 2usize]; 0usize];
     pub static mut RabbitizerRegister_R4000AllegrexM4x4_Names:
         [[*const core::ffi::c_char; 2usize]; 0usize];
+    pub static mut RabbitizerRegister_R4000AllegrexVfpuControl_Names:
+        [[*const core::ffi::c_char; 2usize]; 0usize];
     pub static mut RabbitizerRegister_R5900VF_Names: [[*const core::ffi::c_char; 2usize]; 0usize];
     pub static mut RabbitizerRegister_R5900VI_Names: [[*const core::ffi::c_char; 2usize]; 0usize];
 
@@ -79,6 +81,8 @@ extern "C" {
     pub static mut RabbitizerRegister_R4000AllegrexM2x2_Descriptors: [RegisterDescriptor; 0usize];
     pub static mut RabbitizerRegister_R4000AllegrexM3x3_Descriptors: [RegisterDescriptor; 0usize];
     pub static mut RabbitizerRegister_R4000AllegrexM4x4_Descriptors: [RegisterDescriptor; 0usize];
+    pub static mut RabbitizerRegister_R4000AllegrexVfpuControl_Descriptors:
+        [RegisterDescriptor; 0usize];
 
     /* R5900 */
 
@@ -462,6 +466,29 @@ impl registers_enum::registers::R4000AllegrexM4x4 {
 
         unsafe {
             RabbitizerRegister_R4000AllegrexM4x4_Descriptors.get_unchecked(reg_value as usize)
+        }
+    }
+}
+
+impl registers_enum::registers::R4000AllegrexVfpuControl {
+    pub fn name(self) -> &'static str {
+        let reg_value: u32 = self.into();
+
+        unsafe {
+            std::ffi::CStr::from_ptr(
+                RabbitizerRegister_R4000AllegrexVfpuControl_Names[reg_value as usize][1],
+            )
+            .to_str()
+            .unwrap()
+        }
+    }
+
+    pub fn descriptor(&self) -> &RegisterDescriptor {
+        let reg_value: u32 = (*self).into();
+
+        unsafe {
+            RabbitizerRegister_R4000AllegrexVfpuControl_Descriptors
+                .get_unchecked(reg_value as usize)
         }
     }
 }