+++ src/common/RabbitizerConfig.c
@@ -26,6 +26,7 @@ RabbitizerConfig RabbitizerConfig_Cfg = {
         .userFpcCsr = true,
         .vr4300Cop0NamedRegisters = true,
         .vr4300RspCop0NamedRegisters = true,
+        .r4000AllegrexVfpuControlNamedRegisters = true,
     },
     .pseudos = {
         .enablePseudos = true,