+++ src/instructions/RabbitizerInstrDescriptor.c
@@ -286,6 +286,7 @@ bool RabbitizerInstrDescriptor_hasOperandAlias(const RabbitizerInstrDescriptor *
         case RAB_OPERAND_r4000allegrex_q_vt:
         case RAB_OPERAND_r4000allegrex_q_vd:
         case RAB_OPERAND_r4000allegrex_q_vt_imm:
+        case RAB_OPERAND_r4000allegrex_cop2cd_128:
             break;
 
         case RAB_OPERAND_r4000allegrex_pos: