+++ src/instructions/RabbitizerInstruction/RabbitizerInstruction.c
@@ -390,6 +390,10 @@ void RabbitizerInstruction_blankOut(RabbitizerInstruction *self) {
                 self->word = RAB_INSTR_R4000ALLEGREX_PACK_vt_6_imm(self->word, 0);
                 break;
 
+            case RAB_OPERAND_r4000allegrex_cop2cd_128:
+                self->word = RAB_INSTR_R4000ALLEGREX_PACK_cop2cd_128(self->word, 0);
+                break;
+
             case RAB_OPERAND_r4000allegrex_pos:
                 self->word = RAB_INSTR_R4000ALLEGREX_PACK_pos(self->word, 0);
                 break;