+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- The global `regNames.r4000AllegrexVfpuControlNamedRegisters` option controls
+  if named registers should be used for the VFPU control registers of the R4000
+  ALLEGREX.
+
 ### Changed
 
 - Cleanups in tests code.